refactor: reorganize VectordB cleanup - move legacy/backup scripts to archive
